requirements.txt
langgraph
langchain
langchain-community
langchain-openai
pymupdf
pandas
streamlit
fpdf2